[ #101 ] First step: introduce separate module
