+++ src/syntax/core/mod.rs
@@ -1,9 +1,11 @@
 use super::common::{SyntaxInfo, ToSyntaxInfo};
 
 pub use self::ast::*;
+pub use self::level::*;
 pub use self::pretty::*;
 
 mod ast;
+mod level;
 mod pretty;
 
 impl Val {
